python_scripts/requirements.txt
absl-py==2.3.1
anyio==4.9.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
chex==0.1.90
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
dadapy @ git+https://github.com/sissa-data-science/DADApy@9255cb3772fb93830fe7e823e2e36ab43ce4b5e2
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
etils==1.13.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
flax==0.8.5
fonttools==4.58.4
fqdn==1.5.1
fsspec==2025.5.1
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.12.3
idna==3.10
importlib_resources==6.5.2
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.8
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
ml_dtypes==0.5.3
mpi4py==4.1.0
mpmath==1.3.0
msgpack==1.1.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.3
notebook_shim==0.2.4
numpy==1.26.4
opt_einsum==3.4.0
optax==0.2.5
orbax-checkpoint==0.6.4
overrides==7.7.0
packaging==25.0
pandas==2.3.0
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
protobuf==6.32.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.3
pytest==8.4.0
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.1.0
rpds-py==0.25.1
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
statsmodels==0.14.4
sympy==1.14.0
tensorstore==0.1.76
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
toolz==1.0.0
torch==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20250516
typing_extensions==4.14.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.4.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
zipp==3.23.0
